:art: moved error dialog into ExcelImportNwk.vue
