+++ praktikumsplaner-frontend/frontend/src/components/ExcelImportNwk.vue
@@ -1,54 +1,64 @@
 <template>
-    <v-dialog
-        v-model="visible"
-        persistent
-        max-width="550"
-    >
-        <template #activator="{ on }">
-            <v-btn
-                color="primary"
-                v-on="on"
-            >
-                <v-icon>mdi-tray-arrow-up</v-icon>
-                Datei Hochladen
-            </v-btn>
-        </template>
-        <v-form ref="form">
-            <v-card>
-                <v-card-title class="text-h5 font-weight-bold"
-                    >Datei hochladen</v-card-title
+    <div>
+        <v-dialog
+            v-model="visible"
+            persistent
+            max-width="550"
+        >
+            <template #activator="{ on }">
+                <v-btn
+                    color="primary"
+                    v-on="on"
                 >
-                <v-list>
-                    <v-list-item>
-                        <v-file-input
-                            v-model="excelDatei"
-                            :accept="excelFormat"
-                            :rules="rules"
-                            label="Datei auswählen"
-                            prepend-icon="mdi-tray-arrow-up"
+                    <v-icon>mdi-tray-arrow-up</v-icon>
+                    Datei Hochladen
+                </v-btn>
+            </template>
+            <v-form ref="form">
+                <v-card>
+                    <v-card-title class="text-h5 font-weight-bold"
+                        >Datei hochladen</v-card-title
+                    >
+                    <v-list>
+                        <v-list-item>
+                            <v-file-input
+                                v-model="excelDatei"
+                                :accept="excelFormat"
+                                :rules="rules"
+                                label="Datei auswählen"
+                                prepend-icon="mdi-tray-arrow-up"
+                            >
+                            </v-file-input>
+                        </v-list-item>
+                    </v-list>
+                    <v-card-actions>
+                        <v-spacer />
+                        <v-btn
+                            color="primary"
+                            outlined
+                            @click="cancel()"
                         >
-                        </v-file-input>
-                    </v-list-item>
-                </v-list>
-                <v-card-actions>
-                    <v-spacer />
-                    <v-btn
-                        color="primary"
-                        outlined
-                        @click="cancel()"
-                    >
-                        Abbrechen
-                    </v-btn>
-                    <v-btn
-                        color="primary"
-                        @click="uploadFile()"
-                    >
-                        Hochladen
-                    </v-btn>
-                </v-card-actions>
-            </v-card>
-        </v-form>
-    </v-dialog>
+                            Abbrechen
+                        </v-btn>
+                        <v-btn
+                            color="primary"
+                            @click="uploadFile()"
+                        >
+                            Hochladen
+                        </v-btn>
+                    </v-card-actions>
+                </v-card>
+            </v-form>
+        </v-dialog>
+        <Error-dialog
+            :dialogtext="errorDialogText"
+            :dialogtitle="errorDialogTitle"
+            icontext="mdi mdi-alert-octagon-outline"
+            iconcolor="red"
+            :value="errorDialog"
+            @close="errorDialog = false"
+        ></Error-dialog>
+    </div>
 </template>
 
 <script setup lang="ts">
@@ -57,7 +67,8 @@ import { useSnackbarStore } from "@/stores/snackbar";
 import { Levels } from "@/api/error";
 import NwkService from "@/api/NwkService";
 import { useRules } from "@/composables/rules";
-import { EventBus } from "@/stores/event-bus";
+import ErrorDialog from "@/components/common/ErrorDialog.vue";
+
 const visible = ref<boolean>();
 const excelDatei = ref<File>();
 const form = ref<HTMLFormElement>();
@@ -74,6 +85,12 @@ const rules = [
         "Falsches Dateiformat. Es muss eine Excel-Datei hochgeladen werden."
     ),
 ];
+const errorDialog = ref<boolean>(false);
+const errorDialogText = ref<string>(
+    "Ihre Exceldatei konnte nicht hochgeladen werde. Bitte überprüfen Sie die Datei und versuchen Sie es erneut."
+);
+const errorDialogTitle = ref<string>("Excel Import fehlgeschlagen");
+
 function cancel() {
     visible.value = false;
     form.value?.reset();
@@ -88,11 +105,16 @@ function uploadFile() {
             })
         )
         .catch(() => {
-            EventBus.$emit("excelUploadError");
+            showError();
         })
         .finally(() => {
             cancel();
         });
+}
+
+function showError() {
+    visible.value = false;
+    errorDialog.value = true;
 }
 </script>
